API Strategy for Business Models
APIs form a core part of many organization’s digital strategies, both for companies founded with a technology focus, along with those at the early stages of learning to leverage technology. In this post, we explore what an API is, why they are so valuable to modern business, and how organizations can use APIs to expand and enhance their product offerings and monetize them to drive additional revenue.